Excel Project
This is an excel data cleaning and visualization project. Data used comprised demographics of individual and whether they purchased a bike or not. 
Duplicated data was removed, some variables were replaced , and new columns created using if statements to group  features. 
Finally, an interactive dashboard is provided for further analysis.